Add files via upload
When testing this, I got an Error 403 from AWS, which is a good sign because it means these files can reach AWS. I wasn't sure how we'll set up the CLI wihtin the github; I figure we'll cross that bridge when we get there. I also plan to update this later with the locals.tf file in that same tutorial to make this code easier to update. I just wanted to get something out there for now